+++ src/main/java/web/models/User.java
@@ -9,7 +9,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "surname")
@@ -21,18 +21,18 @@ public class User {
 
     }
 
-    public User(int id, String name, String surname, int salary) {
+    public User(Long id, String name, String surname, int salary) {
         this.id = id;
         this.name = name;
         this.surname = surname;
         this.salary = salary;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
